+++ src/main/java/com/cells/parts/AbstractCombinedInterfacePart.java
@@ -48,12 +48,16 @@ import appeng.tile.inventory.AppEngInternalInventory;
 import appeng.util.SettingsFrom;
 import appeng.util.inv.InvOperation;
 
+import com.cells.api.IInterfaceHost;
+import com.cells.api.IInterfaceProvider;
+import com.cells.api.IUpgradeable;
 import com.cells.blocks.combinedinterface.ICombinedInterfaceHost;
 import com.cells.blocks.interfacebase.AbstractResourceInterfaceLogic;
 import com.cells.blocks.interfacebase.IInterfaceLogic;
 import com.cells.blocks.interfacebase.fluid.FluidInterfaceLogic;
 import com.cells.blocks.interfacebase.item.ItemInterfaceLogic;
 import com.cells.gui.CellsGuiHandler;
+import com.cells.helpers.InterfaceApiHelper;
 import com.cells.integration.mekanismenergistics.CombinedInterfaceGasHelper;
 import com.cells.integration.mekanismenergistics.MekanismEnergisticsIntegration;
 import com.cells.integration.thaumicenergistics.CombinedInterfaceEssentiaHelper;
@@ -70,7 +74,8 @@ import com.cells.network.sync.ResourceType;
  */
 public abstract class AbstractCombinedInterfacePart extends PartBasicState
         implements IGridTickable, ICombinedInterfaceHost,
-                   ItemInterfaceLogic.Host, FluidInterfaceLogic.Host {
+           ItemInterfaceLogic.Host, FluidInterfaceLogic.Host,
+           IInterfaceProvider, IUpgradeable {
 
     protected final IActionSource actionSource;
 
@@ -137,6 +142,18 @@ public abstract class AbstractCombinedInterfacePart extends PartBasicState
     @Override @Nullable public IInterfaceLogic getGasLogic() { return this.gasLogic; }
     @Override @Nullable public IInterfaceLogic getEssentiaLogic() { return this.essentiaLogic; }
     @Override public List<IInterfaceLogic> getAllLogics() { return this.allLogics; }
+
+    @Override
+    @Nonnull
+    public List<IInterfaceHost> getInterfaceHosts() {
+        return InterfaceApiHelper.createInterfaceHosts(this, getTargetFacings());
+    }
+
+    @Override
+    @Nonnull
+    public AppEngInternalInventory getUpgradeInventory() {
+        return this.itemLogic.getUpgradeInventory();
+    }
 
     @Override
     @Nullable